refactor(web-awesome): split VirtualTree into separate folder and components
