+++ packages/web-awesome/src/components/Tree/styles.scss
@@ -230,10 +230,3 @@
   }
 }
 
-.virtual-tree-container {
-  flex: 1 1 auto;
-  min-height: 0;
-  overflow-y: auto;
-  scrollbar-width: thin;
-  height: 100%;
-}
