+++ packages/web-awesome/src/components/VirtualTree/index.tsx
@@ -1,0 +1,1 @@
+export { VirtualTreeList } from "./VirtualTreeList";
